requirements.txt
PySide6
ultralytics
torch
opencv-python
PyYAML
